+++ cmake/ReSolveFindCudaLibraries.cmake
@@ -10,6 +10,10 @@ target_link_libraries(
                          CUDA::cudart
 )
 
+if(RESOLVE_USE_CUDSS)
+  target_link_libraries(resolve_cuda INTERFACE cudss)
+endif()
+
 if(RESOLVE_USE_PROFILING)
   target_link_libraries(resolve_cuda INTERFACE CUDA::nvToolsExt)
 endif()