+++ resolve/SystemSolver.cpp
@@ -20,6 +20,9 @@
 #include <resolve/LinSolverDirectCuSolverRf.hpp>
 #include <resolve/LinSolverDirectCuSparseILU0.hpp>
 #include <resolve/workspace/LinAlgWorkspaceCUDA.hpp>
+#ifdef RESOLVE_USE_CUDSS
+#include <resolve/LinSolverDirectCuDssRf.hpp>
+#endif
 #endif
 
 #ifdef RESOLVE_USE_HIP
@@ -316,6 +319,12 @@ namespace ReSolve
     else if (refactorizationMethod_ == "cusolverrf")
     {
       refactorizationSolver_ = new ReSolve::LinSolverDirectCuSolverRf();
+#ifdef RESOLVE_USE_CUDSS
+    }
+    else if (refactorizationMethod_ == "cudssrf")
+    {
+      refactorizationSolver_ = new ReSolve::LinSolverDirectCuDssRf();
+#endif
 #endif
 #ifdef RESOLVE_USE_HIP
     }
@@ -452,7 +461,11 @@ namespace ReSolve
       return factorizationSolver_->refactorize();
     }
 
-    if (refactorizationMethod_ == "glu" || refactorizationMethod_ == "cusolverrf" || refactorizationMethod_ == "rocsolverrf")
+    if (refactorizationMethod_ == "glu" || refactorizationMethod_ == "cusolverrf" || refactorizationMethod_ == "rocsolverrf"
+#ifdef RESOLVE_USE_CUDSS
+        || refactorizationMethod_ == "cudssrf"
+#endif
+    )
     {
       is_solve_on_device_ = true;
       return refactorizationSolver_->refactorize();
@@ -499,7 +512,7 @@ namespace ReSolve
       is_solve_on_device_ = true;
       status += refactorizationSolver_->setup(A_, L_, U_, P_, Q_);
     }
-    if (refactorizationMethod_ == "cusolverrf")
+    else if (refactorizationMethod_ == "cusolverrf")
     {
       status += refactorizationSolver_->setup(A_, L_, U_, P_, Q_);
 
@@ -508,6 +521,17 @@ namespace ReSolve
 
       is_solve_on_device_ = false;
     }
+#ifdef RESOLVE_USE_CUDSS
+    else if (refactorizationMethod_ == "cudssrf")
+    {
+      LinSolverDirectCuDssRf* Rf = dynamic_cast<LinSolverDirectCuDssRf*>(refactorizationSolver_);
+      Rf->setNumericalProperties(1e-14, 1e-1);
+
+      status += refactorizationSolver_->setup(A_, L_, U_, P_, Q_);
+
+      is_solve_on_device_ = false;
+    }
+#endif
 #endif
 
 #ifdef RESOLVE_USE_HIP
@@ -564,7 +588,11 @@ namespace ReSolve
       status += factorizationSolver_->solve(rhs, x);
     }
 
-    if (solveMethod_ == "glu" || solveMethod_ == "cusolverrf" || solveMethod_ == "rocsolverrf")
+    if (solveMethod_ == "glu" || solveMethod_ == "cusolverrf" || solveMethod_ == "rocsolverrf"
+#ifdef RESOLVE_USE_CUDSS
+        || solveMethod_ == "cudssrf"
+#endif
+    )
     {
       if (is_solve_on_device_)
       {